backend/requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.5
flask-migrate>=4.0.4
flask-jwt-extended>=4.5.2
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-deepseek>=0.0.1
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0
werkzeug>=2.3.7
gunicorn>=21.2.0
python-multipart>=0.0.6
numpy>=1.24.3
tenacity>=8.2.2
aiohttp>=3.8.5
pgvector>=0.2.0